refactor: move code around
